refactor mcp
